touch-drivers: xiaomi: Keep gestures enabled in nonui mode
Disabling and enabling gestures during suspend usually requires a
touchscreen reset. Since the sensor readings can potentially be
jumpy it is questionable whether disabling the gestures every
time is worthy. Since a reset usually involves some delay and
takes a while get rid of the touchscreen reset for pocket mode
changes to avoid the need for touschreen drivers to be able to
handle gesture changes during suspend at all (though support for
this is still nice to get the gestures working on the first sleep
after boot). Instead, just stop reporting the gestures to
userspace if nonui mode is activated.

Change-Id: I16da70b25218717a69aed66030611768cb2c8c37